New experiments, and other changes.

+++ src/python/borg/tools/reextend.py
@@ -10,8 +10,9 @@ logger = borg.get_logger(__name__, default_level = "INFO")
     root = ("root path for search", "positional"),
     old_suffix = ("old file suffix", "positional"),
     new_suffix = ("new file suffix", "positional"),
+    pretend = ("new file suffix", "flag"),
     )
-def main(root, old_suffix, new_suffix):
+def main(root, old_suffix, new_suffix, pretend = False):
     """Change the extensions of multiple files."""
 
     for old_path in borg.util.files_under(root, [old_suffix]):
@@ -21,7 +22,8 @@ def main(root, old_suffix, new_suffix):
 
         logger.info("renaming %s to %s", old_path, new_path)
 
-        os.rename(old_path, new_path)
+        if not pretend:
+            os.rename(old_path, new_path)
 
 if __name__ == "__main__":
     borg.script(main)
